refactor: decouple storage
